% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{Tibble con i dati di pm10 per 4 stazioni della regione VALLEAOSTA}
\format{
Un tibble con 8 colonne e 2924 osservazioni
}
\usage{
pm10
}
\description{
Tibble con i dati di pm10 per 4 stazioni della regione VALLEAOSTA
}
\keyword{datasets}
